requirements_v2.txt
python-dotenv
python-annoy
geopandas
fiona
numpy
scikit-learn
ipython
ipyleaflet
ipykernel
ipywidgets
jupyterlab
joblib
matplotlib
shapely
tqdm
gsutil
pyarrow
xgboost
earthengine-api
geemap
gcsfs
lonboard
python-duckdb
rasterio
